cleanup the code :tada: :sparkles: :tada:
