+++ contact.html
@@ -84,7 +84,7 @@
             </ul>
           </div>
           <div class="contact-form">
-            <form action="/" method="post">
+            <form action="#" method="post">
               <div class="form-group">
                 <label for="first-name">First Name*</label>
                 <input
@@ -240,6 +240,6 @@
       </div>
     </footer>
   </body>
-  <script src="./scripts/script.js" type="module"></script>
+  <script src="./scripts/toggleMenu.js"></script>
   <script src="./scripts/formValidations.js"></script>
 </html>
